// Frontend/MakesCentsFrontend/types/get-year-budget-dto-model.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

import { Month } from "./month";

export class GetYearBudgetDTOModel {
    budgetId: number;
    userId: number;
    month: Month = 1;
    year: number;
    budgetName: string = "";
}